requirements.txt
streamlit
streamlit-authenticator
pyyaml
bcrypt
pandas
psycopg2-binary